feat: module structure
