+++ src/index.html
@@ -38,69 +38,24 @@
         </div>
     </div>
     <div class="content-winners">
-        <p class="number-garage">Winners (4)</p>
+        <p class="number-garage">Winners (<span class="number-winners">1</span>)</p>
         <p class="number-page">page #1</p>
-        <table>
-            <thead>
-                <tr>
-                    <th colspan="3">The table of winners in races</th>
-                </tr>
-            </thead>
-            <tbody>
-                <tr>
-                    <td class="table-number">1.</td>
-                    <td>with two columns</td>
-                    <td>with two columns</td>
-                </tr>
-                <tr>
-                    <td class="table-number">2.</td>
-                    <td>with two columns</td>
-                    <td>with two columns</td>
-                </tr>
-                <tr>
-                    <td class="table-number">3.</td>
-                    <td>with two columns</td>
-                    <td>with two columns</td>
-                </tr>
-                <tr>
-                    <td class="table-number">4.</td>
-                    <td>with two columns</td>
-                    <td>with two columns</td>
-                </tr>
-                <tr>
-                    <td class="table-number">5.</td>
-                    <td>with two columns</td>
-                    <td>with two columns</td>
-                </tr>
-                <tr>
-                    <td class="table-number">6.</td>
-                    <td>with two columns</td>
-                    <td>with two columns</td>
-                </tr>
-                <tr>
-                    <td class="table-number">7.</td>
-                    <td>with two columns</td>
-                    <td>with two columns</td>
-                </tr>
-                <tr>
-                    <td class="table-number">8.</td>
-                    <td>with two columns</td>
-                    <td>with two columns</td>
-                </tr>
-                <tr>
-                    <td class="table-number">9.</td>
-                    <td>with two columns</td>
-                    <td>with two columns</td>
-                </tr>
-                <tr>
-                    <td class="table-number">10.</td>
-                    <td>with two columns</td>
-                    <td>with two columns</td>
-                </tr>
-            </tbody>
-        </table>
+            <table>
+                <thead>
+                    <td>N</td>
+                    <td>COLOR</td>
+                    <td>NAME</td>
+                    <td>WINS</td>
+                    <td>TIME</td>
+                </thead>
+                <tbody class="table-container"></tbody>
+            </table>
+            <div class="nav-bottom">
+                <button>Prev page</button>
+                <button>Next page</button>
+            </div>
+        </div>
     </div>
-   
     <div class="modal-window">
         <div class="modal-text"></div>
         <button class="button-modal-close">close window</button>
