+++ webpack.config.js
@@ -2,7 +2,6 @@ const path = require('path');
 const { merge } = require('webpack-merge');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const { CleanWebpackPlugin } = require('clean-webpack-plugin');
-const CopyPlugin = require("copy-webpack-plugin");
 
 const baseConfig = {
     entry: path.resolve(__dirname, './src/index.js'),
@@ -18,13 +17,6 @@ const baseConfig = {
                 use: 'ts-loader',
                 exclude: /node_modules/,
             },
-            {
-                test: /.(svg)$/i,
-                type: 'asset/resource',
-                generator: {
-                  filename: 'assets/svg/[name][ext][query]'
-                },
-              },
         ],
     },
     resolve: {
@@ -40,12 +32,6 @@ const baseConfig = {
             filename: 'index.html',
         }),
         new CleanWebpackPlugin(),
-        new CopyPlugin({
-            patterns: [
-                { from: "src/assets/svg", to: "assets/svg" }
-                // { from: "other", to: "public" },
-              ]
-        })
     ],
 };
